+++ WebApplication1/Areas/Client/Controllers/ReceivdProposalsController.cs
@@ -62,7 +62,7 @@ namespace WebApplication1.Areas.Client.Controllers
             {
                 return HttpNotFound();
             }
-            PropAndUserVM.ProposalData = _db.Proposals.Where(p => p.IdOfProposal == id);
+            PropAndUserVM.ProposalData = _db.Proposals.Where(p => p.IdOfProposal == id && p.IsAccepted == null);
             var m = from p in _db.Users
                     join u in PropAndUserVM.ProposalData
                     on p.Id equals u.FreeLancerId
@@ -76,30 +76,29 @@ namespace WebApplication1.Areas.Client.Controllers
         [ValidateAntiForgeryToken]
         public async Task<ActionResult> Approve(int id)
         {
-            var proposalDb = await _db.Proposals.FirstOrDefaultAsync(p=>p.IdOfProposal == id);
-            
+            var proposalDb = await _db.Proposals.FirstOrDefaultAsync(p => p.IdOfProposal == id);
+
             if (proposalDb == null)
             {
                 return HttpNotFound();
             }
-            
-            try
-            {
-                proposalDb.IsAccepted = true;
+            proposalDb.IsAccepted = true;
 
-                var postid = proposalDb.PostId;
-                var AllProposelWithTheId = await _db.Proposals.FirstOrDefaultAsync(p => p.PostId == postid && p.IsAccepted == null);
-                var PostOfTheID = await _db.PostJobs.FirstOrDefaultAsync(p => p.Id == postid);
-                PostOfTheID.IsAvilavbleAtWall = false;
+            var postid = proposalDb.PostId;
+            await _db.SaveChangesAsync();
+
+            var AllProposelWithTheId = await _db.Proposals.FirstOrDefaultAsync(p => p.PostId == postid && p.IsAccepted == null);
+            var PostOfTheID = await _db.PostJobs.FirstOrDefaultAsync(p => p.Id == postid);
+            PostOfTheID.IsAvilavbleAtWall = false;
+
+            await _db.SaveChangesAsync();
+            if (AllProposelWithTheId != null)
+            {
                 _db.Proposals.Remove(AllProposelWithTheId);
-
                 await _db.SaveChangesAsync();
-                return RedirectToAction("Index");
             }
-            catch
-            {
-                return View();
-            }
+
+            return RedirectToAction("Index");
         }
 
         // GET: Client/ReceivdProposals/Reject/5
